+++ jax/experimental/fastar.py
